refactoring of zone equip/remove, shield ref update on requip actions
