+++ src/Perpetuum.RequestHandlers/Zone/Containers/RemoveModule.cs
@@ -1,68 +1,17 @@
-using System.Collections.Generic;
-using System.Transactions;
-using Perpetuum.Data;
+using Perpetuum.Containers;
 using Perpetuum.Host.Requests;
-using Perpetuum.Items;
-using Perpetuum.Zones;
+using Perpetuum.Players;
 
 namespace Perpetuum.RequestHandlers.Zone.Containers
 {
-    public class RemoveModule : ZoneContainerRequestHandler
+    public class RemoveModule : ZoneChangeModule
     {
-        public override void HandleRequest(IZoneRequest request)
+        public override void DoChange(IZoneRequest request, Player player, Container container)
         {
-            using (var scope = Db.CreateTransaction())
-            {
-                var character = request.Session.Character;
-                var player = request.Zone.GetPlayerOrThrow(character);
-                //Store ratios before transaction
-                var hpRatio = player.ArmorPercentage;
-                var apRatio = player.CorePercentage;
-
-                CheckPvpState(player).ThrowIfError();
-                CheckCombatState(player).ThrowIfError();
-                CheckActiveModules(player).ThrowIfError();
-
-                var containerEid = request.Data.GetOrDefault<long>(k.containerEID);
-                var container = request.Zone.FindContainerOrThrow(player, containerEid);
-
-                CheckContainerType(container).ThrowIfError();
-                CheckFieldTerminalRange(player, container).ThrowIfError();
-
-                player.EnlistTransaction();
-                container.EnlistTransaction();
-
-                Transaction.Current.OnCompleted(completed =>
-                {
-                    player.Initialize(character);
-
-                    //Apply original ratios after equip-action
-                    player.Armor = player.ArmorMax * hpRatio;
-                    player.Core = player.CoreMax * apRatio;
-
-                    player.SendRefreshUnitPacket();
-
-                    var result = new Dictionary<string, object>
-                    {
-                        {k.robot, player.ToDictionary()}, 
-                        {k.container, container.ToDictionary()}
-                    };
-                    Message.Builder.FromRequest(request).WithData(result).WrapToResult().Send();
-                });
-
-                var moduleEid = request.Data.GetOrDefault<long>(k.moduleEID);
-                var module = player.GetModule(moduleEid).ThrowIfNull(ErrorCodes.ModuleNotFound);
-                module.Owner = character.Eid;
-                module.Unequip(container);
-
-                player.Initialize(character);
-                player.CheckEnergySystemAndThrowIfFailed();  
-
-                player.Save();
-                container.Save();
-                
-                scope.Complete();
-            }
+            var moduleEid = request.Data.GetOrDefault<long>(k.moduleEID);
+            var module = player.GetModule(moduleEid).ThrowIfNull(ErrorCodes.ModuleNotFound);
+            module.Owner = player.Character.Eid;
+            module.Unequip(container);
         }
     }
 }
